#65: Add AchievementService unit tests.
